requirements.txt
jupyter==1.0.0
matplotlib==3.8.2
pandas==2.1.3
scikit-learn==1.3.2
seaborn==0.13.0
folium==0.14.0
bleach~=6.2.0
beautifulsoup4~=4.12.3
soupsieve~=2.6
h11~=0.14.0
defusedxml~=0.7.1
numpy~=1.26.4
pillow~=11.0.0
pip~=23.2.1
attrs~=24.2.0
pyzmq~=26.2.0
pexpect~=4.9.0
tornado~=6.4.2
cffi~=1.17.1
setuptools~=75.6.0
comm~=0.2.2
traitlets~=5.14.3
fqdn~=1.5.1
parso~=0.8.4
jedi~=0.19.2
pytz~=2024.2
packaging~=24.2
QtPy~=2.4.2
PyYAML~=6.0.2
sniffio~=1.3.1
arrow~=1.3.0
python-dateutil~=2.9.0.post0
babel~=2.16.0
Pygments~=2.18.0
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.0
json5~=0.10.0
overrides~=7.7.0
charset-normalizer~=3.4.0
threadpoolctl~=3.5.0
platformdirs~=4.3.6
argon2-cffi-bindings~=21.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
Jinja2~=3.1.4
branca~=0.8.0
requests~=2.32.3
MarkupSafe~=3.0.2
joblib~=1.4.2
psutil~=6.1.0
certifi~=2024.8.30
debugpy~=1.8.9
decorator~=5.1.1
stack-data~=0.6.3
matplotlib-inline~=0.1.7
jupyter_client~=8.6.3
prompt_toolkit~=3.0.48
wcwidth~=0.2.13
ptyprocess~=0.7.0
cycler~=0.12.1
ipywidgets~=8.1.5
urllib3~=2.2.3
six~=1.16.0
anyio~=4.6.2.post1
nbformat~=5.10.4
jupyter_core~=5.7.2
nbclient~=0.10.0
fastjsonschema~=2.21.0
jsonschema~=4.23.0
jupyter_server~=2.14.2
jupyterlab~=4.2.6
jupyterlab_server~=2.27.3
notebook_shim~=0.2.4
notebook~=7.2.2
asttokens~=2.4.1
contourpy~=1.3.1
executing~=2.1.0
fonttools~=4.55.0
ipykernel~=6.29.5
nbconvert~=7.16.4
pandocfilters~=1.5.1
mistune~=3.0.2
jupyterlab_pygments~=0.3.0
pure_eval~=0.2.3
pycparser~=2.22
pyparsing~=3.2.0
qtconsole~=5.6.1
websocket-client~=1.8.0
terminado~=0.18.1
referencing~=0.35.1
webcolors~=24.11.1
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
rfc3339-validator~=0.1.4
jsonschema-specifications~=2024.10.1
async-lru~=2.0.4
kiwisolver~=1.4.7
Send2Trash~=1.8.3
jupyter-lsp~=2.2.5
jupyter-events~=0.10.0
python-json-logger~=2.0.7
prometheus_client~=0.21.0
jupyter_server_terminals~=0.5.3
jupyter-console~=6.6.3
uvicorn~=0.32.1
fastapi~=0.115.5
pydantic~=2.5.2
pytest~=8.3.3